+++ src/iterator/index.ts
@@ -1,0 +1,12 @@
+function map () {
+
+}
+
+
+function reduce () {
+
+}
+
+function filter () {
+  
+}