+++ src/operator/curry.ts
@@ -1,6 +1,7 @@
 export function curry(fn: Function, ...args: any[]){
   return (..._args: any[]) =>
-    (
-      rest => rest.length >= fn.length ? fn(...rest) : curry(fn, rest)
+    (rest => rest.length >= fn.length 
+      ? fn(...rest) 
+      : curry(fn, rest)
     )([ ...args, ..._args ])
 }